+++ tlm_adjoint/firedrake/block_system.py
@@ -78,7 +78,8 @@ where `u_0`, `u_1`, and `u_2` are :class:`firedrake.function.Function` or
 representations of a mixed space solution.
 """
 
-from ..block_system import MixedSpace, iter_sub, zip_sub
+from ..block_system import (
+    BlockNullspace, MixedSpace, NoneNullspace, Nullspace, iter_sub, zip_sub)
 
 from firedrake import (
     Constant, DirichletBC, Function, TestFunction, assemble)
@@ -124,156 +125,6 @@ def apply_bcs(u, bcs):
         u_bc = u
     for bc in bcs:
         bc.apply(u_bc)
-
-
-class Nullspace(ABC):
-    """Represents a matrix nullspace and left nullspace.
-    """
-
-    @abstractmethod
-    def apply_nullspace_transformation_lhs_right(self, x):
-        r"""Apply the nullspace transformation associated with a matrix action
-        on :math:`x`,
-
-        .. math::
-
-            x \rightarrow (I - V (V^* C V)^{-1} V^* C) x.
-
-        :arg x: Defines :math:`x`.
-        """
-
-        raise NotImplementedError
-
-    @abstractmethod
-    def apply_nullspace_transformation_lhs_left(self, y):
-        r"""Apply the left nullspace transformation associated with a matrix
-        action,
-
-        .. math::
-
-            y \rightarrow (I - M U (U^* M U)^{-1} U^*) y.
-
-        :arg y: Defines :math:`y`.
-        """
-
-        raise NotImplementedError
-
-    @abstractmethod
-    def constraint_correct_lhs(self, x, y):
-        r"""Add the linear constraint term to :math:`y`,
-
-        .. math::
-
-            y \rightarrow y + M U S V^* C x.
-
-        :arg x: Defines :math:`x`.
-        :arg y: Defines :math:`y`.
-        """
-
-        raise NotImplementedError
-
-    @abstractmethod
-    def pc_constraint_correct_soln(self, u, b):
-        r"""Add the preconditioner linear constraint term to :math:`u`,
-
-        .. math::
-
-            u \rightarrow u + V \tilde{S}^{-1} U^* b,
-
-        with
-
-        .. math::
-
-            \tilde{S}^{-1} =
-                \left( V^* C V \right)^{-1}
-                S^{-1}
-                \left( U^* M U \right)^{-1}.
-
-        :arg u: Defines :math:`u`.
-        :arg b: Defines :math:`b`.
-        """
-
-        raise NotImplementedError
-
-    def correct_soln(self, x):
-        """Correct the linear system solution so that it is orthogonal to
-        space spanned by the columns of :math:`V`.
-
-        :arg x: The linear system solution, to be corrected.
-        """
-
-        self.apply_nullspace_transformation_lhs_right(x)
-
-    def pre_mult_correct_lhs(self, x):
-        """Apply the pre-left-multiplication nullspace transformation.
-
-        :arg x: Defines the vector on which the matrix action is computed.
-        """
-
-        self.apply_nullspace_transformation_lhs_right(x)
-
-    def post_mult_correct_lhs(self, x, y):
-        """Apply the post-left-multiplication nullspace transformation, and add
-        the linear constraint term.
-
-        :arg x: Defines the vector on which the matrix action is computed, and
-            used to add the linear constraint term. If `None` is supplied then
-            the linear constraint term is not added.
-        :arg y: Defines the result of the matrix action on `x`.
-        """
-
-        self.apply_nullspace_transformation_lhs_left(y)
-        if x is not None:
-            self.constraint_correct_lhs(x, y)
-
-    def correct_rhs(self, b):
-        """Correct the linear system right-hand-side so that it is orthogonal
-        to the space spanned by the columns of :math:`U`.
-
-        :arg b: The linear system right-hand-side, to be corrected.
-        """
-
-        self.apply_nullspace_transformation_lhs_left(b)
-
-    def pc_pre_mult_correct(self, b):
-        """Apply the pre-preconditioner-application nullspace transformation.
-
-        :arg b: Defines the vector on which the preconditioner action is
-            computed.
-        """
-
-        self.apply_nullspace_transformation_lhs_left(b)
-
-    def pc_post_mult_correct(self, u, b):
-        """Apply the post-preconditioner-application left nullspace
-        transformation, and add the linear constraint term.
-
-        :arg u: Defines the result of the preconditioner action on `b`.
-        :arg b: Defines the vector on which the preconditioner action is
-            computed, and used to add the linear constraint term. If `None` is
-            supplied then the linear constraint term is not added.
-        """
-
-        self.apply_nullspace_transformation_lhs_right(u)
-        if b is not None:
-            self.pc_constraint_correct_soln(u, b)
-
-
-class NoneNullspace(Nullspace):
-    """An empty nullspace and left nullspace.
-    """
-
-    def apply_nullspace_transformation_lhs_right(self, x):
-        pass
-
-    def apply_nullspace_transformation_lhs_left(self, y):
-        pass
-
-    def constraint_correct_lhs(self, x, y):
-        pass
-
-    def pc_constraint_correct_soln(self, u, b):
-        pass
 
 
 class ConstantNullspace(Nullspace):
@@ -414,70 +265,6 @@ class DirichletBCNullspace(Nullspace):
 
     def pc_constraint_correct_soln(self, u, b):
         self._constraint_correct_lhs(b, u, alpha=1.0 / self._alpha)
-
-
-class BlockNullspace(Nullspace):
-    """Nullspaces for a mixed space.
-
-    :arg nullspaces: A :class:`.Nullspace` or a :class:`Sequence` of
-        :class:`.Nullspace` objects defining the nullspace. `None` indicates a
-        :class:`.NoneNullspace`.
-    """
-
-    def __init__(self, nullspaces):
-        if not isinstance(nullspaces, Sequence):
-            nullspaces = (nullspaces,)
-
-        nullspaces = list(nullspaces)
-        for i, nullspace in enumerate(nullspaces):
-            if nullspace is None:
-                nullspaces[i] = NoneNullspace()
-        nullspaces = tuple(nullspaces)
-
-        super().__init__()
-        self._nullspaces = nullspaces
-
-    def __new__(cls, nullspaces, *args, **kwargs):
-        if not isinstance(nullspaces, Sequence):
-            nullspaces = (nullspaces,)
-        for nullspace in nullspaces:
-            if nullspace is not None \
-                    and not isinstance(nullspace, NoneNullspace):
-                break
-        else:
-            return NoneNullspace()
-        return super().__new__(cls)
-
-    def __getitem__(self, key):
-        return self._nullspaces[key]
-
-    def __iter__(self):
-        yield from self._nullspaces
-
-    def __len__(self):
-        return len(self._nullspaces)
-
-    def apply_nullspace_transformation_lhs_right(self, x):
-        assert len(self._nullspaces) == len(x)
-        for nullspace, x_i in zip(self._nullspaces, x):
-            nullspace.apply_nullspace_transformation_lhs_right(x_i)
-
-    def apply_nullspace_transformation_lhs_left(self, y):
-        assert len(self._nullspaces) == len(y)
-        for nullspace, y_i in zip(self._nullspaces, y):
-            nullspace.apply_nullspace_transformation_lhs_left(y_i)
-
-    def constraint_correct_lhs(self, x, y):
-        assert len(self._nullspaces) == len(x)
-        assert len(self._nullspaces) == len(y)
-        for nullspace, x_i, y_i in zip(self._nullspaces, x, y):
-            nullspace.constraint_correct_lhs(x_i, y_i)
-
-    def pc_constraint_correct_soln(self, u, b):
-        assert len(self._nullspaces) == len(u)
-        assert len(self._nullspaces) == len(b)
-        for nullspace, u_i, b_i in zip(self._nullspaces, u, b):
-            nullspace.pc_constraint_correct_soln(u_i, b_i)
 
 
 class Matrix(ABC):
