+++ tlm_adjoint/firedrake/hessian_system.py
@@ -3,12 +3,12 @@ from ..interface import (
     var_axpy, var_axpy_conjugate, var_copy, var_copy_conjugate, var_dtype,
     var_inner, var_space, var_space_type)
 
-from ..block_system import MixedSpace, iter_sub, tuple_sub
+from ..block_system import (
+    BlockNullspace, MixedSpace, NoneNullspace, iter_sub, tuple_sub)
 from ..eigendecomposition import eigendecompose
 from ..manager import manager_disabled
 
-from .block_system import (
-    BlockNullspace, Matrix, NoneNullspace, Preconditioner, System)
+from .block_system import Matrix, Preconditioner, System
 
 from collections.abc import Sequence
 import numpy as np
